pyproject.toml
[project]
name = "viirsnow"
version = "0.1.0"
description = "Code for paper \"Evaluation of VIIRS snow cover products in French mountains using Sentinel-2\""
authors = [
    {name = "Nicola Imperatoren", email = "nicola.imperatore@meteo.fr"},
]
dependencies = [
    "ipykernel>=6.30.1",
    "numpy>=2.3.3",
    "xarray>=2025.9.0",
    "dask>=2025.9.1",
    "netcdf4>=1.7.2",
    "matplotlib>=3.10.6",
    "scores>=2.2.0",
    "geopandas>=1.1.1",
    "rioxarray>=0.19.0",
    "geospatial_grid @ git+ssh://git@github.com/nicolaimpe/geospatial_grid@0.1.0",
    "mountain_data_binner @ git+ssh://git@github.com/nicolaimpe/mountain_data_binner@0.1.0",
    "ndsi_fsc_calibration @ git+ssh://git@github.com/nicolaimpe/ndsi_fsc_calibration@0.1.0",
    "seaborn>=0.13.2",
    "jinja2>=3.1.6",
]

requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
